Feat(Game): create heat manager, interior volume, door script
Door should open and close, open doors deny you that sweet +2 bonus to warmth for being inside, making you rely more on the fireplace.

I'm going to make the player slow down with frostbite next, so we'll see where this all ends up after a bit of testing.

Current refresh rate on calculating and dealing frostbite to player is set to 5x per second. if that's too fast or slow we can adjust the max frostbite to kill or the refresh rate itself. As is- you have to be out in -2 heat for 100 seconds to die.
-1 heat for 200 seconds to die.  I'll probably end up tweaking it with some playtesting. Make it easy enough to actually get to day 4 at least.